Speed up get_assets_for_run_id when there are a lot of events
Summary:
Current impl of this method gets every log and filters them down for events that happen to contain asset materializations. This is potentially very expensive if there are a lot of non-system logs though. Instead, use our index to filter down to events with a particular event type.

I think to make this method truly efficient we are going to need a method on run_id + event_type though...