Add a "log framework error" hook to execute_step invocations, similar to the one in execute_run (#6690)
Summary:
Today we had an issue where a call to execute_step failed in system code, and the event log was left blank for that step. This PR attempts to make it so that framework errors like that are still surfaced in the event log and associated with that step.