RFC: Don't require an @repository to be defined when there is only one repository - gather all code artifacts and package them together instead
Summary:
Inspired by https://threads.com/34418826592 from nick - if we are planning to enforce that each process has a single repository, and are generally planning to de-emphasize the repository as a core organization thing in Dagit, we have the option of not requiring you to write an @repository at all. Instead we could bring in each of the code artifacts in the file/module and package them up into one for you. This might be a bit more familiar to people coming from airflow's auto-geneated DAG approach too.

Whipped this up to explore that direction.

One big thing we would need to figure out is caching - the existing code loading paths rely on the fact that there's a single RepositoryDefinition in code to apply some caching, so creating a new one each time the code is loaded may be undesirable.