Also override cpu/memory in containerOverrides
Previously, we only set our cpu/memory overrides as a taskOverride.

This sets the limit for the entire task. If the containers don't
have their own limit set, they'll expand to consume all available
cpu/memory in the task. This is the case when we rely on Dagster to
register its own task definitions because the task definitions it
registers don't set container cpu/memory.

But if the containers have their own limits defined, they'll continue to
respect that limit even if we override the task itself. This can happen
when you provide a custom task definition that explicitly sets
cpu/memory for containers.

This fix changes the behavior of the EcsRunLauncher to set cpu/memory
overrides as both a taskOverride and a containerOverride. This way, both
the task will be upsized and the container will be allowed to consume
cpu/memory beyond its initial allocation.

One small gotcha in the PR that I've called out in a comment - AWS
expects that cpu/memory be an integer in containerOverrides but a string
in taskOverrides.