Don't carry forward all container definitions (#6850)
When we create a task definition for the job run, we inherit a bunch of
information from the container running running the current process. The
original implementation assumed that this was the lone container running
in the task (or at the very least, that the other containers had no
meaningful side effects).

But if somebody chooses to run a task with both a daemon and a dagit
container (or even a handful of sidecar containers), this behavior would
cause the dagit container to also stand up in the job run task.

This change constrains that behavior to only launch a single container
in the job run task.