Add short-circuit in run start time if run failed to start (#6991)
Summary:
If a run fails to start while launching, it gets a FAILURE event but no STARTED event. In that case, it currently falls back to the expensive get_run_stats query. Instead, fall back to the end time as the same as the start time.

Test Plan: Use a run launcher that fails on startup, load the runs page