Don't reuse task definitions if secrets change
Another thing we can't override is the secrets used by a container:

https://github.com/aws/containers-roadmap/issues/1269

So before reusing the task definition, check to see that the secrets in
our container definition are the same as the secrets in the upcoming
run. Keep in mind that adding/removing secrets with the "dagster" tag in
Secrets Manager could cause this to change between runs even without a
change to dagster.yaml.