fix logic for turning schedule ticks into hour boundaries when there's an offset (#7071)
Summary:
We were subtracting the hours to reach midnight, which works great most of the time, but does not work so great at 7AM on a spring DST morning (where going back 7 hours sends you to 11PM the previous night since you skipped an hour). Fix that and add more test coverage of daily partitions with hour offets.