Optionally use sidecars (#7087)
First, we changed the EcsRunLauncher to only carry forward a single
container definition:

https://github.com/dagster-io/dagster/pull/6850

But this change didn't account for container definitions that defined
sidecars with `dependsOn`. When using the docker compose CLI to deploy
ECS, this introduced a breaking change because docker compose runs each
of its ECS tasks with a dependent sidecar container.

We fixed this with:

https://github.com/dagster-io/dagster/pull/6929

But it was only a partial fix; it removed the ability to intentionally
provide sidecars while still relying on Dagster to generate an ECS task
definition for you.

This change gives users the option to continue to create sidecars by
setting `use_sidecars` to `True` in their dagster.yaml. Perhaps in
the future, we'll change this config to optionally allow a list of
specific sidecars to enable (for example, if you have a monitoring
sidecar that you don't want the run task to inherit but a secrets
sidecar that you do want it to inherit). But for now, this should allow
users to configure their deployments to continue to behave the way
things behaved before 6850.