Don't give up if the terminate() call raises an exception during force-termination
Summary:
The idea with force-terminate is that no matter what the run is moved into a terminal state, but if an exception is raised we give up. this fixes that by logging to the event log if there's an exception, but still moving ahead and force-terminating the run.

Test Plan: BK