Rename the InstanceMigratedError to be less pushy about migrating
Summary:
A lot of the logic here is from a time when every migration was required - but today lots of migrations are not, and the actual error is just as interesting as whether the alembic revision is out of date.

Rename the error and the language to be more like a warning instead of a "you had better migrate right now or else" error.

Test Plan: BK, manually trigger a SQL error while the db is out of date and view the resulting stack trace in dagit