[dagit] Fix config refresh when reloading repo (#7325)
## Summary

Resolves #6312.

When setting config editor sessions for jobs with preset configs, we aren't currently setting the preset as the "base" of the session. This means that when the repository is reloaded, we don't notice that the session's config may need to be refreshed as well, so we never prompt the user.

To fix this, set the preset in the session `base`.

## Test Plan

Follow repro steps in attached issue. Verify that when I modify the job config and reload the repository in Dagit, I am prompted to refresh the config. Verify that refreshing the config works properly.