[dagit] Fix CodeMirror show-hint error (#7334)
## Summary

There is an error within the CodeMirror show-hint addon that fires on an empty Launchpad. This appears to be because we aren't providing `from` and `to` values in the return value for the `hint` helper function, and they are expected to be there. Add them.

## Test Plan

View Launchpad, clear out editor config. Start typing in the empty editor, verify that hint suggestions render properly and that there are no JS errors in the console.