[dagit] Allow toggling repos in bulk (#7344)
## Summary

Resolves #7083.

Add a checkbox to the `RepoSelector` to allow selecting or unselecting the entire list of repos.

This involves moving the "X of Y selected" text to the `RepoSelector` itself, as well as changing the `onToggle` function to accept an array.

## Test Plan

Load Dagit with seven repos loaded, open repo selector. Verify correct behavior of top-level checkbox and individual checkboxes.