Remove lazy load example from repository docblock
Summary:
This feature isn't really compatible with CRAG - for example, schedules require passing in the job object now, not just the job name, so there isn't any feasible path to using these lazy-loaded objects as soon as you have a schedule or sensor.

We don't need to remove the underlying functionality or anything, but we shouldn't expose it in the docblock if it will stop working as soon as you use a schedule.