Allow jobs to be passed in lazily to repository definitions (#7382)
Summary:
This check was crashing when a job was passed in lazily.

It will still crash when a graph is passed in lazily (the assumption that repos consist of jobs not graphs appears to be pretty deep) but this will make it so that users can pass in a lambda that returns a job. The error is a bit odd, but is the same error you would get if you passed in a job function that returned some other random object that wasn't a job.