make k8s container context volumes and volume mounts use snake case, not camel case
Summary:
I realized when writing the docs that this field sticks out like a sore thumb in the typeahead since it's the only thing that expects snake case. We already have logic that is set up to coerce both snake case and camel case to be the right form, but this makes the default config that container_context uses use snake case instead of camel case.

Test Plan: BK