Show repo selector in unloadable schedules/sensors instead of repo location metadata
Summary:
When you change a repo name and there's now an unloadable schedule, we let you see information about it in the Schedules page.

This logic used to be a lot more finnicky before phil's migration last week, so it was useful to show you all the various bits of metadata that might have changed and made it unloadable (e.g. the module name, the python executable, that kind of thing).

Now, after phil's migration, the only thing that will make a schedule unloadable is changing the repo name or the schedule name. So let's show that instead in the metadata. Also, since there's only one thing, let's ditch the "show info" button and just always show it. (The visual treatment here could probably be better)

This has the nice side effect of making this "show info" button not blank in Cloud as well (where we don't actually have access to information about the schedule's origin, just the selector)