Unbreak stopping unloadable schedules/sensors (#7446)
Summary:
The new stop mutations assume that the ExternalSensor can be loaded, which is not the case for unloadable schedules and sensors. Getting this right requires threading the selector ID all the way through into the mutation path (at least in the unloadable path, but might as well do it everywhere for consistency)

Will add a test at the graphql python level as well.