Fix back-compat for new helm chart param for container context
Summary:
I incorrectly thought that the default values.yaml of False would be applied for this param, but since it's in a list of deployments, the values in values.yaml do not matter. Fix that and add tests that still use the old format where includeConfigInLaunchedRuns is not set at all.

Test Plan: BK