[graph] use the correct input definition for type checks (#7453)
https://dagster.phacility.com/D6155 added things to the input source types to allow for removing all user code from the execution plan structure. 

`input_name` and `solid_handle` values were added to the source objects. The way these values are populated is that they correspond to where along the mapping chain the dependency was resolved. This is important for the `FromConfig` since the config may be supplied at different places in the config structure which map to different input mapping layers.
Unfortunately, for other sources such as `FromStepOutput` this information is more confusing than useful. 

This subtle confusion was fine with `@composite_solid` since it enforced that all types were equal along a mapping chain.

`@graph` dropped this requirement, which exposes this subtle "bug" and caused type checks to source the `dagster_type` from the mapping input def, which in the `@graph` case is likely `Any` instead of the `@op` input def. 


This diff fixes the underlying issue by passing `input_def` directly and deprecating the confusing properties from the input sources they are not needed on.


resolves #7452

## Test Plan

test added

still need to ensure that the serdes changes are safe and do not require more extensive back-compat / forward-compat protections .