Register a new task definition for IAM changes (#7564)
We no longer register a new task definition for every run:

https://github.com/dagster-io/dagster/pull/7009

This introduced a regression where changing the IAM role used by a task
definition doesn't re-register the task definition. So it tries to run
the task with an IAM role that doesn't exist. The easiest way to
reproduce this is to use our example `docker compose` functionality:

https://github.com/dagster-io/dagster/issues/7358

This change extends our logic for deciding when to reuse a task
definition. Now, we reuse a task definition if:

- The container definition's image, name, or secrets change
- The task definition's task role or execution role changes

For now, I've pulled the logic out into its own private method with a
separate test so we can more easily extend the logic. I think
eventually, this would benefit from a broader refactor - perhaps with
the introduction of a `TaskDefinition` class that implements `__eq__`.