Use explicitly set env vars rather than autoenvvar prefixes for the grpc command
Summary:
We make use of this kind of buried click feature that auto-creates env vars for you, but I feel weird recommending that people set the DAGSTER_CLI_API_GRPC_CONTAINER_CONTEXT env var to use container context - it's a weird name, and its not really greppable. This gives us a more natural env var name to recommend.

I added a "GRPC" prefix for the options that are specifically about the server (host/socket/port) as opposed to the code loaded by the server where I just went with DAGSTER_XXX.