[dagit] Show default value for config (#7598)
## Summary

Expose `defaultValueAsJson` as a field on `CompositeConfigType`, and use it to display config defaults in the launchpad schema pane.

For fields in the schema that do have defaults, render a subtle underline. Hover on the field, and a tooltip appears with the default config value.

I don't know if there's a good way to expose defaults without hiding them behind some interaction, because the space here is so limited.

Open to others' thoughts on this.

## Test Plan

View Launchpad, hover over keys in the schema pane. Verify that the tooltip appears, with nicely formatted JSON.