[RFC] lighter weight pull request template (#7538)
The existing template was created when we were internally using phabricator, so the target audience was only external contributors.

Now that we are using GH, I feel the template should strike a better balance between having enough information for external contributors but also being stream lined for us internally.

Personally, this was motiviated by seeing internal PRs that submit without filling out the template (or editing it away).

Very open to feedback on this, just wanted to start a discussion.