[dagit] Better indentation guide in our CodeMirrors (#7765)
### Summary & Motivation

Resolves #3006.

Replace whitespace markers with indentation guides in our CodeMirror usage. This is a longstanding request for improving the UX of the editor experience.

- Guides are rainbowed to make it a bit clearer which indentation level is which. I tried to alternate the contrast a little bit too.
- The existing whitespace toggle has been removed. IMO the indentation guides are strictly better than having nothing at all, so I don't think we need to bother making it toggleable.

<img width="336" alt="Screen Shot 2022-05-06 at 10 48 09 AM" src="https://user-images.githubusercontent.com/2823852/167168925-4da3aa12-4e51-4082-abae-325f3fbb9f34.png">

(Note that in this screenshot, I have aded newlines to demonstrate the guide colors. They don't stretch all the way to the bottom of the editor window.)

### How I Tested These Changes

View Launchpad editor, add a bunch of content to test out indentation levels. Verify proper rendering of guides.

View instance config and post-run config (read-only CodeMirrors), verify same.