Allow TextInput to have a type of number (#7778)
### Summary & Motivation

Sometimes we need a number input. I didn't see one in the codebase, the easiest option would be to allow a type of `number` to the existing `TextInput`. A number is technically a subset of text  🤔. We could alternatively create a separate NumberInput component that shares a Base with the current TextInput component in case we want to treat them differently in the future. I thought it might be overkill to do that right now though

### How I Tested These Changes
I rendered an input with a type of "number" for a cloud PR