[dagit] Fix global search for repeated item keys (#7827)
### Summary & Motivation

There is currently a bug where users who have jobs/sensors/schedules that share the same name end up with a broken global search experience because keys are repeated.

To repair this, use `href`, which should be guaranteed unique for various objects, even if they have the same name.

### How I Tested These Changes

View Dagit as a user with repeated object names for jobs/sensors. Open global search, verify that rendering and behavior (typing, key navigation) is correct when searching for a name that would have previously collided.