Don't call it a 'framework error' when a run worker unexpectedly restarts
Summary:
A user pointed out that it's confusing if Dagster times out a run because it took too long to start, then it eventually starts up, then Dagster confusingly claims it was a framework error. Change that specific case to be more of an FYI instead of a framework error.

There's also a good case to be made I think that we shouldn't treat the existing case when this happens when the run was already started as a 'framework error' either (although we should certainly still fail the run if the cluster interrupts a run and then tries to start it up again from the beginning)

Test Plan: Bk