RFC: Better error handling when an ExternalRepository grpc call fails
Summary:
I can't come up with a case where this *should* happen - to the point where I'm having trouble finding a good test to write for this (an error getting raised not when the repo is first loaded, but when we go to fetch the external repository), but at least one user-submitted issue suggests that it *does* happen: https://github.com/dagster-io/dagster/issues/7927. Add better error handling like the other gRPC methods for this where we will surface a nice stack trace when it does.

This is technically a minor breaking change, in that errors on older versions that used to fail with one gross exception (DagsterUserCodeUnreachableError with an inscrutable stack trace) will now fail with a different gross exception until you upgrade (failed to find ExternalRepositoryErrorData to deserialize it). If we wanted to be really sneaky we could repurpose a class like  ExternalSensorExecutionErrorData, which has the same arguments, to avoid that.