Error rather than warn when an event log query comes in that tries to use an int cursor on an event log DB that needs a RunShardedEventsCursor (#7970)
Summary:
Right now, if you pass in an int cursor to a SqliteEventLogStorage query, we warn and then ignore the cursor. The warning incorrectly implies that it will just be slow, but actually the cursor isn't applied at all, which is very unlikely to be what the caller expects. This PR moves that case to an exception instead.

If we're concerned about this being a breaking change we could wait until 0.15.0.