[dagit] Fix SearchBootstrapQuery to be lazy (#7981)
### Summary & Motivation

Make the global search dialog's bootstrap query be lazy (upon opening the search dialog) instead of loading it up front.

I had thought this was already lazy, but apparently not.

### How I Tested These Changes

Load Dagit, verify that no `SearchBootstrapQuery` is fired. Click "Search", verify that it is fired and that search behaves normally.