Remove can_terminate from run launchers and run coordinators, use a status check instead in graphql (#7983)
Summary:
The default implementation of this when the run is running involves a DB query for each run that you call it on (and in the backfill UI we might call it on lots of running runs). I claim that this method is no longer needed and in the vast majority of cases a status check is sufficient. It also brings open source dagster in line with Dagster Cloud, which just uses a status check for termination.

Test Plan: BK, terminate a run from dagit