WIP: Don't fetch every single run in the InstanceBackfillsQuery (#7985)
Summary:
This elminates places where the (old) backfill ui loads every single run in every backfill. The problem doesn't appear to be the fact that we need to fetch all the runs, but rather the gigantic graphql response when we return 10s of thousands of runs to the client.

I did leave a 'cancelableRuns' field that I would love to remove as well - that appears to mainly exist so that Dagit can do buik termination from the client. Instead we should write a 'terminate every run in backfill X' bulk action and not need to send thousands of run IDs to dagit, but that is future work after this PR.

For now i just took the new backill UI out of the existing path to test this out. I won't land that - prha has graciously volunteered to split the two UIs a bit more before we land this.

I also still need to add tests for the new fields I added.

However, this drops a large backfill page with ~16k runs load time from about 4 minutes to about 15 seconds.