Fix issue where we added the run ID repeatedly to the runs for the tick during scheduler failure recovery
Summary:
If you fail when launching a run and then retry, we previously would keep appending the run ID to the end of the list, causing a bunch of duplicate run IDs to appear in the run view on the Schedule page. Now, check to verify that it's in there before adding.

Test Plan: BK (additional test logic now passes, failed before during test_failure_recovery_between_multi_runs)