Handle InvalidSubsetError in launch run results
Summary:
A user reported that a launch run mutation that triggered an InvalidSubsetError was raised as an unhandled graphql exception. Add that explicitly to the set of errors that are handled in the graphql query and client-side.

Not sure how I feel about this java-y "catch all exceptions that might possibly be thrown from the server" pattern though.

There's also a visual display issue with this showCustomAlert path when there's only a message but no stack trace - i'll post a screenshot. To repro swap in this implementation on the server:

```
@capture_error
def launch_pipeline_execution(graphene_info, execution_params):
    from ...schema.errors import GrapheneInvalidSubsetError
    from ...schema.pipelines.pipeline import GraphenePipeline
    from ..utils import UserFacingGraphQLError, capture_error

    raise UserFacingGraphQLError(
        GrapheneInvalidSubsetError(
            message="sup",
            pipeline=GraphenePipeline(
                graphene_info.context.get_full_external_pipeline(execution_params.selector)
            ),
        ),
    )
    return _launch_pipeline_execution(graphene_info, execution_params)
```

Test Plan: Simulate this error being thrown from the server using the error above, launch run - no more uncaught graphql exception.