Load all code artifacts on grpc server startup, not just pipelines/jobs
Summary:
User reported a race condition with partition set loading - we have logic in place to prevent this for jobs/pipelines, but not for the other code artifacts. Fix that by adding a new method that pulls in any lazy code artifacts at grpc server startup time.

Test Plan:
Add log outputs to validate that partition sets are loaded as soon as the grpc server is started up