[dagit] Enable sectioned left nav by default (#8039)
### Summary & Motivation

Make the sectioned left nav opt-out. This just flips the logic of the feature flag, and shows the User Settings flag as an opt-out with a flipped boolean.

### How I Tested These Changes

View User Settings, verify that the nav is sectioned and that the flag switch is on by default. Turn it off, verify that the nav and the switch are correctly toggled. Reload, verify that flag state has been saved.