Skip flaky snowflake+pandas tests
Summary:
This test that I added that runs the snowflake IO manager is failing with errors like:

```
[2022-05-25T22:04:30Z] df =    foo  quux
[2022-05-25T22:04:30Z] 0  bar     1
[2022-05-25T22:04:30Z] 1  baz     2
[2022-05-25T22:04:30Z] 2  bar     1
[2022-05-25T22:04:30Z] 3  baz     2
[2022-05-25T22:04:30Z]
[2022-05-25T22:04:30Z]     @op
[2022-05-25T22:04:30Z]     def read_pandas_df(df: pandas.DataFrame):
[2022-05-25T22:04:30Z]         assert set(df.columns) == {"foo", "quux"}
[2022-05-25T22:04:30Z] >       assert len(df.index) == 2
[2022-05-25T22:04:30Z] E       assert 4 == 2
[2022-05-25T22:04:30Z] E         +4
[2022-05-25T22:04:30Z] E         -2
```

This is presumably because multiple tests are running the same job in parallel, so are writing to the snowflake table simultaneously.

disable it to make master green again, but i'm curious - does this indicate a real problem with the IO manager? generally I would expect that I would be able to run multiple jobs with the same IO manager in parallel (but I am specifying a specific table here, so I see how that is tricky here)