Fix upstream context handling in fs_asset_io_manager (#8007)
* Add test case that is currently failing

* Add helper function to get asset identifier from InputContext

* Rework asset path determination logic in fs_asset_io_manager

By passing either InputContext or OutputContext to _get_path, we can
implement it without using upstream_output for assets.

Using upstream_output for assets was problematic, because fetching the
partition_key required upstream step config, which is not available in
upstream_output.

* Refactor to enable change of arguments in get_output_context.

* Emmit warning on use of step_context through upstream_output

This change should be reverted once upstream_output.step_context is
initialized as None in PlanExecutionContext.for_input_manager

* Add uniform get_asset_identifier for both InputContext and OutputContext

* Add uniform get_identifier method for InputContext and OutputContext

* Adapt to new API

* Fix a couple of typing issues related to optionals

* Fix another typing issue

Black formatting unsilenced a mypy error that was explicitly ignored. We
move things around to fix the original mypy complain.

* Change all uses of get_output_identifier to get_identifier

This fixes an error on dagster-tests.