Make dagster-images build work when it is called outside of a git repo (#8088)
Summary:
Right now there's a requirement that when you run this command you have the automation library checked out in a github repo. The release pipeline will break that requirement once it is running from a Docker image. So instead figure out where the code is to copy based on the current github repository, rather than assuming a fixed file relative to the python file running the command.

Test Plan:
BK