Attempt to speed up query for existing run keys for a given sensor
Summary:
DB profiling suggests that the existing query here is quite expensive. Instead of doing a join across sensor name

The way that this could go wrong is if a user is sharing run key formats across lots of sensors - but I think that is less likely to happen than the current situation, and there are mitigations we can apply if we observe that.