Use status column for reading runs from sql storage (#8528)
Use the status in the status column of run rows, rather than whatever is serialized in the run body

This makes it harder for concurrent writes to a run- e.g. updating status and updating tags- to overwrite eachother.

This has test coverage by the handle_run_event tests, a little tricky to write more

Cloud will follow