Parse incoming config into YAML on the client rather than in python on the server
Summary:
This is mostly a workaround for a weird pyyaml issue where it interprets numeric strings as octal - since we are representing the config as dictionaries on the server, it seems reasonable to me to expose them that way in the graphql API as well? (With the caveat that GenericScalars are weird / kind of a graphql anti-pattern).

Test Plan:
Very that config with numeric strings:

["0001234", "0001235", "0001236", "0001237"]

is rendered correctly in the launchpad and still launches correctly

Verify that job re-execution and config editing and local run config storage still works