Better support for nested causes in dagit
Summary:
Right now if there's an exception chain that's 3 exceptions deep, we don't show the juicy inner exception in dagit (which is often the most useful part), because of the way that our error type is structured. This PR moves it to return a list of causes instead.

Test Plan: New BK, view similar error in dagit