Don't hit the gRPC server at all during asset backfill or AMP ticks
Summary:
This is a prototype of a way to run AMP and asset backfills without going through the gRPC server at all (other than to initially generate the workspace and asset graph that is used by the tick).

Today the way that we interact with user code in these systems is we:
- Load the workspace / asset graph (from the DB in cloud)
- Once we decide what runs to create, we make a gRPC call to generate a job snapshot and an execution plan snapshot for each run (possibly running newer code that is not compatible with what is in the DB)
- Enqueue the run - it then runs in a container image based on what is in the DB, but is using a snapshot that was potentially generated on newer code

If we land this PR, we will instead use the code in the DB for everything and never make any gRPC calls. We do this by:
- Not worrying about the job snapshot at all and just not storing it on the run, I have not found anything yet that breaks from doing this for asset runs of the type produced by AMP and asset backfills
- Waiting to generate the execution plan snapshot until the run starts, relying on the fact that it will be running using the same image as opposed to newer, potentially incompatible code. The only user-visible result of this change is that you don't see the gantt chart in the UI until the run is dequeued.