dagster project scaffold: move dagit and pytest to extras_require 1/2 (#9273)
### Summary & Motivation
as part of dogfooding to deploy examples to cloud, when we are installing the example as a python package, i.e. using setup.py to install deps, i realized dagit and pytest aren't deps of the project itself and dagit is a heavy dep - they are more of a "dev" dependencies. so i moved them to "dev" and updated the readme and doc to refer to `pip install ".[dev]"` (alternatively, we could get rid of them, but then in the readme, we'll need to mention `pip install dagit`. i feel this pr's path is easier for new users.

### How I Tested These Changes
scaffolded a new project. `pip install -e ".[dev]"` worked