[dagit] Refetch queries instead of dumping Apollo cache (#9406)
### Summary & Motivation

Now that we have `refetchQueries` available in Apollo, use it instead of resetting the entire Apollo store. This will hopefully make it a bit less jarring when the user reloads a repository or entire workspace.

### How I Tested These Changes

Load Dagit. Reload a repo after making changes to its Python, verify with console logging that queries are refetched and that the UI refreshes correctly. Reload the entire Workspace, verify same.