[dagit] Add elapsed time tooltips to log row timestamps (#9479)
### Summary & Motivation

Resolves #9209.

Add a tooltip to log timestamps to show the elapsed time since the start of the run, and if applicable, the elapsed time since the start of the relevant step.

I did some refactoring of the time utilities here. Tests are included.

<img width="367" alt="Screen Shot 2022-08-24 at 2 00 27 PM" src="https://user-images.githubusercontent.com/2823852/186552897-a89b527d-1dc3-42b9-99c6-b1e7493fd9dd.png">


### How I Tested These Changes

View a run, hover over timestamps. Verify that the tooltips render correctly, including "negative" times that are for logs that occur prior to the start of the run.