[dagit] [workspace] Create Jobs tab under Workspace (#9472)
### Summary & Motivation

Create a "Jobs" tab under Workspace that includes a repo-bucketed list of all jobs in the Workspace, with table rendering similar to the current Instance Overview page.

This change also includes the addition of `WorkspaceTabs` and a new `JobsTable` used both here and in Instance Overview, and some changes to the structure of job item objects used in the table.

### How I Tested These Changes

Enable feature flag, view Status and Workspace pages. Verify that they render the correct pages, tabs, etc.

Disable feature flag, verify that rendering and behavior is unchanged from current UI.

Tested in Cloud as well, which will need a handful of changes for custom routes.