[dagit] Fix zero-index reference on potentially empty array (#9541)
### Summary & Motivation

Based on an error that appeared in Cloud.

We aren't checking whether `assetMaterializations` actually has a zero element, just that `asset` is not nullish. The result is that checking `timestamp` on the zero-index element can lead to a runtime error.

Ensure that the array has a length before looking at its first element.

### How I Tested These Changes

Buildkite. Sanity check that an asset view loads.