docs mdxcomponents: change TabGrouop syntax to work with remark format DREL-363 (#9610)
### Summary & Motivation
the problem with snapshot or format cmd is that the mdx formatter (i.e. remarkjs) falsely identifies the `<TabGroup ... />` as `text` when the component gets longer.
i didn't find a good way to tell the formatter that it should be `html`.

so this pr changes the `TabGroup` syntax to adapt to the formatter, i.e. make the html components shorter.

### How I Tested These Changes
`make mdx-format` doesn't change the `TabGroup` component in mdx file